Wrote a bunch of HACKS that allowed me to do problem e and g.
Basically, put the whole nu thing in logspace to compute faster, had it trapezoidally integrate (1 - line depth) for each zeta, and then return those "equivalent widths" in both frequency and wavelength space.

I made the plots for e and g interactively instead of in a nice function. I want to go to sleep.